Create intial project starting point
The project now can be worked on. There are a lot of new stuff so I will give
a brief overview of them here and explain them more in the pull request.

Src directory-
Where the source code is. It is currently just running some tests.

Scripts-
There is a bat script in there to build the project.

Vendor-
There are a bunch of directories for the dependencies.

premake5.lua-
This is the script for the build system.