ImGui (WIP 2)
Moved bits of code around and populated the ImGui_ImplVulkan_InitInfo struct. Added an extra parameter for the Gui in the CreateInfo struct in the Renderer class.

Right now, the code compiles and runs fine, but the demo gui does not display. I'm working on debugging that at the moment, but I want to merge what I have so far so others can work on it.